Refactor example package structure
Select some examples to sit at the top of the package structure
Add package-info
