+++ examples/src/main/java/com/opengamma/strata/examples/finance/package-info.java
@@ -1,0 +1,14 @@
+/**
+ * Copyright (C) 2016 - present by OpenGamma Inc. and the OpenGamma group of companies
+ *
+ * Please see distribution for license.
+ */
+
+/**
+ * More detailed examples of Strata.
+ * <p>
+ * This package contains the more detailed examples of using Strata.
+ * The parent package, {@code com.opengamma.strata.examples}, has higher-level examples.
+ */
+package com.opengamma.strata.examples.finance;
+
