+++ examples/src/main/java/com/opengamma/strata/examples/gui/package-info.java
@@ -1,0 +1,13 @@
+/**
+ * Copyright (C) 2016 - present by OpenGamma Inc. and the OpenGamma group of companies
+ *
+ * Please see distribution for license.
+ */
+
+/**
+ * GUI views of Strata.
+ * <p>
+ * This package contains an example GUI designed to assist with schedule generation.
+ */
+package com.opengamma.strata.examples.gui;
+
