+++ examples/src/main/java/com/opengamma/strata/examples/marketdata/credit/markit/package-info.java
@@ -1,0 +1,11 @@
+/**
+ * Copyright (C) 2016 - present by OpenGamma Inc. and the OpenGamma group of companies
+ *
+ * Please see distribution for license.
+ */
+
+/**
+ * Credit market data for examples.
+ */
+package com.opengamma.strata.examples.marketdata.credit.markit;
+
