+++ examples/src/main/java/com/opengamma/strata/examples/marketdata/package-info.java
@@ -1,0 +1,11 @@
+/**
+ * Copyright (C) 2016 - present by OpenGamma Inc. and the OpenGamma group of companies
+ *
+ * Please see distribution for license.
+ */
+
+/**
+ * Market data for examples.
+ */
+package com.opengamma.strata.examples.marketdata;
+
