+++ examples/src/main/java/com/opengamma/strata/examples/package-info.java
@@ -1,0 +1,15 @@
+/**
+ * Copyright (C) 2016 - present by OpenGamma Inc. and the OpenGamma group of companies
+ *
+ * Please see distribution for license.
+ */
+
+/**
+ * High level examples of Strata.
+ * <p>
+ * This package contains the high-level examples of using Strata.
+ * More examples can be found in {@code com.opengamma.strata.examples.finance},
+ * including examples of calibration, scenarios and trade construction.
+ */
+package com.opengamma.strata.examples;
+
