+++ examples/src/main/java/com/opengamma/strata/examples/report/package-info.java
@@ -1,0 +1,11 @@
+/**
+ * Copyright (C) 2016 - present by OpenGamma Inc. and the OpenGamma group of companies
+ *
+ * Please see distribution for license.
+ */
+
+/**
+ * Integration of examples with strata-report.
+ */
+package com.opengamma.strata.examples.report;
+
